New management action: run action on all remote accesspoints
Example:

  python3 manage.py run_remote_action \
      --command 'ln -s ../tmp/database.json /www/opennet-ondataservice-export.json' \
      --password-file router-passwords.secret \
      --success-file=example-success-hosts.list \
      192.168.1.23 192.168.1.50

The above will run the given command on any reachable host.
The passwords to be tried are read from the password file (on by line).
The IPs of all successfully connected hosts are stored in the optional
"success file".  They will not be tried again in a repeated operation.
Optionally a number of target addresses can be specified.  Otherwise
*all* known hosts are targetted.

Care should be taken when running the action on *all* hosts.  There
should be compliance tests at the beginning, e.g. "[ -e /etc/banner ]"
in order to execute actions on accesspoints only (not on servers).

Please note, that passwords and private ssh keys are used for
connection attempts.  Thus the given action may also be executed on
servers.